Feat: User authentication added using jwt to verify the user and protect each users personal contact storage

+++ controllers/contactController.js
@@ -1,56 +1,64 @@
-const expressAsyncHandler = require("express-async-handler");
+const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
-
-const getContacts = expressAsyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+//@access private
+const getContacts = asyncHandler(async (req, res) => {
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
-//@desc get individual contact
-//@route GET /api/contacts/:id
-//@access public
-
-const getContact = expressAsyncHandler(async (req, res) => {
-  const contact = await Contact.findById(req.params.id);
-  if (!contact) {
-    res.status(404);
-    throw new Error("Contact not found");
-  }
-  res.status(200).json(contact);
-});
-
-//@desc create contacts
+//@desc Create New contact
 //@route POST /api/contacts
-//@access public
-
-const createContact = expressAsyncHandler(async (req, res) => {
-  console.log("Requested Body", req.body);
+//@access private
+const createContact = asyncHandler(async (req, res) => {
+  console.log("The request body is :", req.body);
   const { name, email, phone } = req.body;
-
+  if (!name || !email || !phone) {
+    res.status(400);
+    throw new Error("All fields are mandatory !");
+  }
   const contact = await Contact.create({
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
-  if (!name || !phone) {
-    res.status(400);
-    throw new Error("Please add both name and phone number");
-  }
+
   res.status(201).json(contact);
 });
 
-//@desc update contacts
-//@route PUT /api/contacts/:id
-//@access public
-
-const updateContact = expressAsyncHandler(async (req, res) => {
+//@desc Get contact
+//@route GET /api/contacts/:id
+//@access private
+const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id !== req.user.id) {
+    res.status(403);
+    throw new Error("User unauthorized for this operation");
+  }
+
+  res.status(200).json(contact);
+});
+
+//@desc Update contact
+//@route PUT /api/contacts/:id
+//@access private
+const updateContact = asyncHandler(async (req, res) => {
+  const contact = await Contact.findById(req.params.id);
+  if (!contact) {
+    res.status(404);
+    throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -58,28 +66,31 @@ const updateContact = expressAsyncHandler(async (req, res) => {
     req.body,
     { new: true }
   );
+
   res.status(200).json(updatedContact);
 });
 
-//@desc delete contact
+//@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
-
-const deleteContact = expressAsyncHandler(async (req, res) => {
+//@access private
+const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
-
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
 });
 
 module.exports = {
   getContacts,
+  createContact,
   getContact,
   updateContact,
   deleteContact,
-  createContact,
 };
